job-application-tracker/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.11
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1